+++ selenium-node-firefox/bin/start-selenium-node-firefox.sh
@@ -3,6 +3,10 @@
 # set -e: exit asap if a command exits with a non-zero status
 set -e
 
+echoerr() {
+  cat <<< "$@" 1>&2;
+}
+
 # Wait for this process dependencies
 timeout --foreground ${WAIT_TIMEOUT} wait-xvfb.sh
 timeout --foreground ${WAIT_TIMEOUT} wait-xmanager.sh
@@ -27,7 +31,37 @@ java ${JAVA_OPTS} \
   -cleanUpCycle ${SEL_CLEANUPCYCLE_MS} \
   -nodePolling ${SEL_NODEPOLLING_MS} \
   ${SELENIUM_NODE_PARAMS} \
-  ${CUSTOM_SELENIUM_NODE_PROXY_PARAMS}
+  ${CUSTOM_SELENIUM_NODE_PROXY_PARAMS} \
+  ${CUSTOM_SELENIUM_NODE_REGISTER_CYCLE} \
+  &
+NODE_PID=$!
+
+function shutdown {
+  echo "-- INFO: Shutting down Firefox NODE gracefully..."
+  kill -SIGINT ${NODE_PID} || true
+  kill -SIGTERM ${NODE_PID} || true
+  kill -SIGKILL ${NODE_PID} || true
+  wait ${NODE_PID}
+  echo "-- INFO: Firefox node shutdown complete."
+  # First stop video recording because it needs some time to flush it
+  supervisorctl -c /etc/supervisor/supervisord.conf stop video-rec || true
+  killall supervisord
+  exit 0
+}
+
+function trappedFn {
+  echo "-- INFO: Trapped SIGTERM/SIGINT on Firefox NODE"
+  shutdown
+}
+# Run function shutdown() when this process a killer signal
+trap trappedFn SIGTERM SIGINT SIGKILL
+
+# tells bash to wait until child processes have exited
+wait ${NODE_PID}
+echo "-- INFO: Passed after wait java Firefox node"
+
+# Always shutdown if the node dies
+shutdown
 
 # Note to double pipe output and keep this process logs add at the end:
 #  2>&1 | tee $SELENIUM_LOG